requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
django-silk==4.3.0
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.13.0
gprof2dot==2021.2.21
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.3
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
